// frontend/src/api/generated/shop/schemas/shipment.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * ShopDjango API
 * API for ShopDjango project
 * OpenAPI spec version: 1.0.0
 */

export interface Shipment {
  readonly id: number;
  /** Order this shipment is for */
  order: number;
  /** Shipping method used */
  shipping_method: number;
  /** Courier company handling the shipment */
  courier: number;
  /** When the package was shipped */
  shipped_at: string;
  /**
   * When the package was delivered
   * @nullable
   */
  delivered_at?: string | null;
  /** Shipping address at time of shipment */
  shipping_address: string;
  /** Timestamp when the record was created */
  readonly created_at: string;
  /** Timestamp when the record was last updated */
  readonly updated_at: string;
}